refactor: rename the data privacy module
rename data privace module.
