+++ .github/workflows/python-publish.yml
@@ -3,7 +3,7 @@
 
 # This workflow uses actions that are not certified by GitHub.
 # They are provided by a third-party and are governed by
-# separate terms of service, privacy policy, and support
+# separate terms of service, data_privacy policy, and support
 # documentation.
 
 name: Upload Python Package
