+++ pilot/model/proxy/llms/__init__.py
@@ -1,5 +1,5 @@
 """
     There are several limitations to privatizing large models: high deployment costs and poor performance.
-    In scenarios where data privacy requirements are relatively low, connecting with commercial large models can enable
+    In scenarios where data data_privacy requirements are relatively low, connecting with commercial large models can enable
     rapid and efficient product implementation with high quality.
 """
